broll_bot/requirements.txt
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.0
playwright>=1.40.0
Pillow>=10.0.0
aiohttp>=3.9.0
requests>=2.31.0
python-dotenv>=1.0.0
